拆分聊天 actor 檔案
